refactor production editor page
